Fix ints stored and saved as floats
Integer values should be written as ints to the tree and not as floats.

+++ KappaAnalysis/src/Consumers/KappaCutFlowTreeConsumer.cc
@@ -6,13 +6,13 @@ void KappaCutFlowTreeConsumer::Init(KappaSettings const& settings)
 {
 	CutFlowTreeConsumer<KappaTypes>::Init(settings);
 
-	this->m_runExtractor = [](event_type const& event, product_type const& product, setting_type const& setting) -> float {
+	this->m_runExtractor = [](event_type const& event, product_type const& product, setting_type const& setting) -> int {
 		return event.m_eventInfo->nRun;
 	};
-	this->m_lumiExtractor = [](event_type const& event, product_type const& product, setting_type const& setting) -> float {
+	this->m_lumiExtractor = [](event_type const& event, product_type const& product, setting_type const& setting) -> int {
 		return event.m_eventInfo->nLumi;
 	};
-	this->m_eventExtractor = [](event_type const& event, product_type const& product, setting_type const& setting) -> float {
+	this->m_eventExtractor = [](event_type const& event, product_type const& product, setting_type const& setting) -> uint64_t {
 		return event.m_eventInfo->nEvent;
 	};
 }
